Move binary dependency management back to CUDA.jl.
